removeUser

+++ src/oncology/db/jpa/JPAUserMaster.java
@@ -88,6 +88,12 @@ public class JPAUserMaster implements UserMaster {
 	}
 
 	@Override
+	public List<User> getUsersList(int id_role) {
+		Query q = em.createNativeQuery("SELECT * FROM users WHERE role LIKE" +id_role, User.class);
+		return (List<User>) q.getResultList();
+	}
+	
+	@Override
 	public User checkPassword(String email, String password) {
 		
 		Query q=null;
@@ -133,13 +139,26 @@ public class JPAUserMaster implements UserMaster {
 	}
 	
 	
-	public void removeUser(String email, String password) {
+	public User removeUser(String email, String password) {
 		
-		em.getTransaction().begin();
-		User u=this.checkPassword(email, password);
-		em.remove(u);
-		em.getTransaction().commit();
-
+		
+		User u=this.getUser(email);
+		List<User> usersList=this.getUsersList(u.getRole().getId());
+		for(int i=0; i<usersList.size(); i++) {
+			if(usersList.get(i).getEmail()==email) {
+				
+				em.getTransaction().begin();
+				u=this.checkPassword(email, password);
+				em.remove(u);
+				em.getTransaction().commit();
+			}else {
+				
+				u=null;
+			}
+			
+		}
+		return u;
+		
 	}
